backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
requests>=2.32.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml[html_clean]>=5.2.0
chromadb>=0.5.5
pillow>=10.4.0
numpy>=1.26.0
httpx>=0.27.0
matplotlib>=3.8.0
